fix: 랭킹 데이터가 불일치 하는 문제 해결

+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/spec/dto/response/RankingResponse.java
@@ -26,7 +26,7 @@ public class RankingResponse {
         private double totalAnalyzeScore;
         private int rankByJobField;
         private int totalUsersCountByJobField;
-        private int rank;
+        private long rank;
         private boolean isBookmarked;
         private int commentsCount;
         private int bookmarksCount;
